update setup & install
