+++ .github/workflows/ci_pkg-install.yml
@@ -49,17 +49,19 @@ jobs:
           pip install -r requirements.txt --find-links https://download.pytorch.org/whl/cpu/torch_stable.html
 
       - name: Install | Uninstall package - archive
+        working-directory: ./dist
         run: |
           # install as archive
-          pip install dist/*.tar.gz
+          pip install *.tar.gz
           cd ..
           python -c "import pytorch_lightning as pl ; print(pl.__version__)"
           pip uninstall -y pytorch-lightning
 
       - name: Install | Uninstall package - wheel
+        working-directory: ./dist
         run: |
           # install as wheel
-          pip install dist/*.whl
+          pip install *.whl
           cd ..
           python -c "import pytorch_lightning as pl ; print(pl.__version__)"
           pip uninstall -y pytorch-lightning
